Issue #1148 - Remove Newtonsoft.Json dependency in GraphQL and move dependent apis into GraphQL.Serialization.Newtonsoft
