+++ src/GraphQL.NewtonsoftJson/NewtonsoftContractResolver.cs
@@ -1,0 +1,29 @@
+using System;
+using System.Reflection;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+
+namespace GraphQL.NewtonsoftJson
+{
+    public class NewtonsoftContractResolver : DefaultContractResolver
+    {
+        private readonly CamelCaseNamingStrategy _camelCase = new CamelCaseNamingStrategy();
+
+        protected override JsonConverter ResolveContractConverter(Type objectType) =>
+            objectType == typeof(ExecutionResult)
+                ? new ExecutionResultJsonConverter()
+                : base.ResolveContractConverter(objectType);
+
+        protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization memberSerialization)
+        {
+            var property = base.CreateProperty(member, memberSerialization);
+
+            if (property.DeclaringType == typeof(Instrumentation.ApolloTrace) || property.DeclaringType?.DeclaringType == typeof(Instrumentation.ApolloTrace))
+            {
+                property.PropertyName = _camelCase.GetPropertyName(member.Name, false);
+            }
+
+            return property;
+        }
+    }
+}
