+++ src/GraphQL.Tests/BasicQueryTestBase.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
+using GraphQL.NewtonsoftJson;
 using GraphQL.Types;
-using GraphQL.Http;
 using GraphQL.Validation;
 using GraphQLParser.Exceptions;
 using Shouldly;
