+++ src/GraphQL.Tests/Bugs/Bug118SpacesInTypeNameTests.cs
@@ -1,3 +1,4 @@
+using GraphQL.NewtonsoftJson;
 using GraphQL.Types;
 using Xunit;
