+++ src/GraphQL.Tests/Bugs/Issue661.cs
@@ -1,10 +1,10 @@
-using GraphQL.Http;
+using System;
+using System.Collections.Generic;
 using GraphQL.Types;
+using GraphQL.NewtonsoftJson;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
-using System;
-using System.Collections.Generic;
 using Xunit;
 
 namespace GraphQL.Tests.Bugs
