+++ src/GraphQL.Tests/Execution/InputConversionTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using GraphQL.NewtonsoftJson;
 using Shouldly;
 using Xunit;
