+++ src/GraphQL.Tests/Instrumentation/ApolloTracingTests.cs
@@ -4,6 +4,7 @@ using Shouldly;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace GraphQL.Tests.Instrumentation
@@ -63,6 +64,33 @@ query {
             }
             paths.Count.ShouldBe(expectedPaths.Count);
             new HashSet<List<object>>(paths).ShouldBe(expectedPaths);
+        }
+
+        [Fact]
+        public async Task serialization_should_have_correct_case()
+        {
+            var trace = new ApolloTrace(new DateTime(2019, 12, 05, 15, 38, 00, DateTimeKind.Utc), 102.5);
+            var expected = @"{
+  ""version"": 1,
+  ""startTime"": ""2019-12-05T15:38:00Z"",
+  ""endTime"": ""2019-12-05T15:38:00.103Z"",
+  ""duration"": 102500000,
+  ""parsing"": {
+    ""startOffset"": 0,
+    ""duration"": 0
+  },
+  ""validation"": {
+    ""startOffset"": 0,
+    ""duration"": 0
+  },
+  ""execution"": {
+    ""resolvers"": []
+  }
+}";
+
+            var result = await Writer.WriteToStringAsync(trace);
+
+            result.ShouldBeCrossPlat(expected);
         }
     }
 }
