+++ src/GraphQL/GraphQL.csproj
@@ -27,7 +27,6 @@
   <ItemGroup>
     <PackageReference Include="GraphQL-Parser" Version="4.2.2" />
     <PackageReference Include="Microsoft.CSharp" Version="4.5.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
     <PackageReference Include="System.Buffers" Version="4.5.0" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.5.0" />
     <PackageReference Include="System.Reactive.Core" Version="4.0.0" />
