+++ src/GraphQL/StringExtensions.cs
@@ -1,5 +1,3 @@
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -15,45 +13,6 @@ namespace GraphQL
         public static string ToFormat(this string format, params object[] args)
         {
             return string.Format(format, args);
-        }
-
-        /// <summary>
-        /// Converts a JSON-formatted string into a dictionary.
-        /// </summary>
-        /// <param name="json">A JSON formatted string.</param>
-        /// <returns>Inputs.</returns>
-        public static Inputs ToInputs(this string json)
-        {
-            var dictionary = json != null ? ToDictionary(json) : null;
-            return dictionary == null
-                ? new Inputs()
-                : new Inputs(dictionary);
-        }
-
-        /// <summary>
-        /// Converts a JSON object into a dictionary.
-        /// </summary>
-        public static Inputs ToInputs(this JObject obj)
-        {
-            var variables = obj?.GetValue() as Dictionary<string, object>
-                            ?? new Dictionary<string, object>();
-            return new Inputs(variables);
-        }
-
-        /// <summary>
-        /// Converts a JSON formatted string into a the dictionary.
-        /// </summary>
-        /// <param name="json">The json.</param>
-        /// <returns>Returns a <c>null</c> if the object cannot be converted into a dictionary.</returns>
-        public static Dictionary<string, object> ToDictionary(this string json)
-        {
-            var values = JsonConvert.DeserializeObject(json,
-                new JsonSerializerSettings
-                {
-                    DateFormatHandling = DateFormatHandling.IsoDateFormat,
-                    DateParseHandling = DateParseHandling.None
-                });
-            return GetValue(values) as Dictionary<string, object>;
         }
 
         /// <summary>
@@ -92,56 +51,6 @@ namespace GraphQL
                 return s;
 
             return newFirstLetter + s.Substring(1);
-        }
-
-        /// <summary>
-        /// Gets the value contained in a JObject, JValue, JProperty, or JArray.
-        /// </summary>
-        /// <param name="value">The object containing the value to extract.</param>
-        /// <remarks>If the value is a recognized type, it is returned unaltered.</remarks>
-        public static object GetValue(this object value)
-        {
-            if (value is JObject objectValue)
-            {
-                var output = new Dictionary<string, object>();
-                foreach (var kvp in objectValue)
-                {
-                    output.Add(kvp.Key, GetValue(kvp.Value));
-                }
-                return output;
-            }
-
-            if (value is JProperty propertyValue)
-            {
-                return new Dictionary<string, object>
-                {
-                    { propertyValue.Name, GetValue(propertyValue.Value) }
-                };
-            }
-
-            if (value is JArray arrayValue)
-            {
-                return arrayValue.Children().Aggregate(new List<object>(), (list, token) =>
-                {
-                    list.Add(GetValue(token));
-                    return list;
-                });
-            }
-
-            if (value is JValue rawValue)
-            {
-                var val = rawValue.Value;
-                if (val is long l)
-                {
-                    if (l >= int.MinValue && l <= int.MaxValue)
-                    {
-                        return (int)l;
-                    }
-                }
-                return val;
-            }
-
-            return value;
         }
     }
 }
